Add support for the GRUB emulator to boot Linux entries using kexec
The grub-emu tool is used as a debugging utility but it could also be used
as a user-space bootloader if there is support to boot an operating system.

The Linux kernel is already able to (re)boot another kernel via the kexec
boot mechanism. So the grub-emu tool could rely on this feature and have a
linux and initrd commands that are used to pass a kernel, initramfs image
and kernel command line parameters to kexec for booting a selected entry.

By default the systemctl kexec option is used so systemd can shutdown all
of the running services before doing a reboot using kexec. But if this is
not present, it fallbacks to executing the kexec user-space tool directly.

Signed-off-by: Michael Chang <mchang@suse.com>
[javierm: add commit message, some fixes and coding style cleanup]
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>

Patch-cc: Michael Chang <mchang@suse.com>
Patch-cc: Peter Jones <pjones@redhat.com>
Patch-cc: Vladimir Serbinenko <phcoder@gmail.com>
Patch-cc: Daniel Kiper <daniel.kiper@oracle.com>