10_linux: generate BLS section even if no kernels are found in /boot
The 10_linux script exits if no kernels are found in the /boot directory.
This makes sense for a non-BLS configuration but doesn't when using BLS.

Add a BLS section regardless if there are kernel images in /boot or not.
That way the grub.cfg can be generated even before a kernel is installed.