10_linux: don't search for OSTree kernels
The 10_linux script used to exit if kernels weren't found in the /boot dir
so the path where OSTree kernels are installed was added to prevent that.

But the script should really add a section to call the blscfg command even
if kernels have not been installed, it doesn't make sense to search these.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>