Revert "Disable multiboot, multiboot2, and linux16 modules on EFI builds."
This reverts commit 155d4e84604 which disabled building the multiboot and
multiboot2 modules on EFI builds. But that made the menu entries created
by the Xen package to stop working since they use the multiboot2 module.

The mentioned commit disabled building the multiboot{,2} modules because
they can be used to bypass the Secure Boot mechanism. But it's enough to
not include these modules in the grub2 EFI binary that's signed, which
is the case already in the grub2 package.

Having them as modules if the user installs the grub2-efi-x64-modules is
a valid use case. And since module loading isn't allowed when Secure Boot
is enabled, it doesn't represent any security threat.

Resolves: rhbz#1703872

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>