util/mkimage: Improve data_size value calculation
According to "Microsoft Portable Executable and Common Object File Format
Specification", the Optional Header SizeOfInitializedData field contains:

"Size of the initialized data section, or the sum of all such sections if
there are multiple data sections".

Make this explicit by adding the GRUB kernel data plus the sum of all the
modules sizes. The ALIGN_UP() is not required by the PE spec, but do just
do it to make sure that there won't be alignment issues.

Signed-off-by: Peter Jones <pjones@redhat.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>