mkimage: Unify more of the PE32 and PE32+ header set-up.
There's quite a bit of code duplication in the code that sets the optional
header for PE32 and PE32+. The two are very similar with the exception of
a few fields that have type grub_uint64_t instead of grub_uint32_t.

Factor out the common code and add a PE_OHDR() macro that simplifies the
set-up and make the code more readable. Always use grub_host_to_target32()
function since the constant values set in the grub_uint64_t fields fit in
a grub_uint32_t.

Signed-off-by: Peter Jones <pjones@redhat.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>
Series-changes: 2
- Explain in commit message why grub_host_to_target32() is always used.
- Fix some code style issues.