efi: Print an error if boot to firmware setup is not supported
The "fwsetup" command is only registered if the firmware supports booting
to the firmware setup UI. But it could be possible that the GRUB config
already contains a "fwsetup" entry, because it was generated in a machine
that has support for this feature.

To prevent users getting a "can't find command `fwsetup`" error if it is
not supported by the firmware, let's just always register the command but
print a more accurate message if the firmware doesn't support this option.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>
Series-cc: Paul Whalen <pwhalen@fedoraproject.org>
Series-cc: Daniel Kiper <daniel.kiper@oracle.com>
Series-cc: Peter Robinson <pbrobinson@gmail.com>
Series-cc: Alberto Ruiz <aruiz@gnome.org>
Cover-letter:
Improve logic to check for fwsetup support
The fwsetup command can be used for rebooting to the firmware setup UI but
this can only be used on machines that have support for this feature.

Curently, this is achieved by checking in the 30_uefi-firmware template if
OsIndicationsSupported is present and has EFI_OS_INDICATIONS_BOOT_TO_FW_UI
set. This is again checked in the efifwsetup module init function and the
fwsetup command is only registered if that condition is meet.

But this logic has a comple of problems, since it assumes that the config
file is always generated in the same machine than the one that's booted.

If that's not the case, any of the following situations could happen:

* A config file generated on a non-EFI machine, will not have the fwsetup
  menu entry even when booting with EFI.
* A config file generated on an EFI machine, will have a fwsetup even when
  is not booted with EFI.
* A config file generated on an EFI machine that supports rebooting to the
  firmware setup, will have the fwsetup even if booting on an EFI machine
  that does not support that feature.

To prevent these, let's change when the check is done. Instead of doing it
at config generation time, do it at runtime with grub_platform == "efi" and
when the command is executed. That way, the 30_uefi-firmware template won't
make any assumption about the firmware support of the machine that's booted.

Best regards,
Javier
END